WIP: Extract service/binance into modules
